+++ dash_blueprint_components/SegmentedControl.py
@@ -5,9 +5,9 @@ from dash.development.base_component import Component, _explicitize_args
 
 class SegmentedControl(Component):
     """A SegmentedControl component.
-A SegmentedControl is a linear collection of buttons which allows a user to choose an option from multiple choices, 
-similar to a Radio group. Compared to the ButtonGroup component, SegmentedControl has affordances to signify a 
-selection UI and a reduced visual weight which is appropriate for forms.
+A SegmentedControl is a linear collection of buttons which allows a user to choose an option from multiple choices,
+ similar to a Radio group. Compared to the ButtonGroup component, SegmentedControl has affordances to signify a
+ selection UI and a reduced visual weight which is appropriate for forms.
 
 Keyword arguments:
 
@@ -19,7 +19,16 @@ Keyword arguments:
     element.
 
 - defaultValue (string; optional):
-    Initial value. Mutually exclusive with value prop.
+    Initial value when using component in uncontrolled mode. If value
+    prop is also defined, this value is ignored. Setting this prop,
+    does not initialize the value prop. Use the value prop if you need
+    to listen or read the value of the component in a callback, i.e.
+    use the value in a State or Input.
+
+- disabled (boolean; default False):
+    If True, the option buttons are non-interactive. The value can
+    still controllable via callback if disabled is True. Default is
+    False.
 
 - fill (boolean; optional):
     Whether the control group should take up the full width of its
@@ -44,16 +53,19 @@ Keyword arguments:
     CSS properties to apply to the root element.
 
 - value (string; optional):
-    Selected value. Mutually exclusive with defaultValue prop."""
+    Selected value. When a value is given to this prop, the
+    defaultValue is ignored. When using the value of this component as
+    a state or input in a callback, use this property instead of
+    defaultValue."""
     _children_props = []
     _base_nodes = ['children']
     _namespace = 'dash_blueprint_components'
     _type = 'SegmentedControl'
     @_explicitize_args
-    def __init__(self, id=Component.UNDEFINED, className=Component.UNDEFINED, defaultValue=Component.UNDEFINED, fill=Component.UNDEFINED, inline=Component.UNDEFINED, intent=Component.UNDEFINED, large=Component.UNDEFINED, options=Component.UNDEFINED, small=Component.UNDEFINED, style=Component.UNDEFINED, value=Component.UNDEFINED, **kwargs):
-        self._prop_names = ['id', 'className', 'defaultValue', 'fill', 'inline', 'intent', 'large', 'options', 'small', 'style', 'value']
+    def __init__(self, id=Component.UNDEFINED, className=Component.UNDEFINED, defaultValue=Component.UNDEFINED, disabled=Component.UNDEFINED, fill=Component.UNDEFINED, inline=Component.UNDEFINED, intent=Component.UNDEFINED, large=Component.UNDEFINED, options=Component.UNDEFINED, small=Component.UNDEFINED, style=Component.UNDEFINED, value=Component.UNDEFINED, **kwargs):
+        self._prop_names = ['id', 'className', 'defaultValue', 'disabled', 'fill', 'inline', 'intent', 'large', 'options', 'small', 'style', 'value']
         self._valid_wildcard_attributes =            []
-        self.available_properties = ['id', 'className', 'defaultValue', 'fill', 'inline', 'intent', 'large', 'options', 'small', 'style', 'value']
+        self.available_properties = ['id', 'className', 'defaultValue', 'disabled', 'fill', 'inline', 'intent', 'large', 'options', 'small', 'style', 'value']
         self.available_wildcard_properties =            []
         _explicit_args = kwargs.pop('_explicit_args')
         _locals = locals()